rays: Move shader_source and shader_program to opengl/
Both are entirely GL-dependent. Update include paths in
shader.cpp and painter.cpp.
