+++ rays/src/opengl/shader_program.h
@@ -1,13 +1,13 @@
 // -*- c++ -*-
 #pragma once
-#ifndef __RAYS_SRC_SHADER_PROGRAM_H__
-#define __RAYS_SRC_SHADER_PROGRAM_H__
+#ifndef __RAYS_SRC_OPENGL_SHADER_PROGRAM_H__
+#define __RAYS_SRC_OPENGL_SHADER_PROGRAM_H__
 
 
 #include <xot/pimpl.h>
 #include "rays/defs.h"
 #include "rays/coord.h"
-#include "opengl/opengl.h"
+#include "opengl.h"
 
 
 namespace Rays
