+++ rays/src/opengl/shader_source.cpp
