+++ rays/src/opengl/shader_source.h
@@ -1,11 +1,11 @@
 // -*- c++ -*-
 #pragma once
-#ifndef __RAYS_SRC_SHADER_SOURCE_H__
-#define __RAYS_SRC_SHADER_SOURCE_H__
+#ifndef __RAYS_SRC_OPENGL_SHADER_SOURCE_H__
+#define __RAYS_SRC_OPENGL_SHADER_SOURCE_H__
 
 
 #include <xot/pimpl.h>
-#include "opengl/opengl.h"
+#include "opengl.h"
 
 
 namespace Rays
