+++ rays/src/painter.cpp
@@ -20,11 +20,11 @@
 #include "image.h"
 #include "font.h"
 #include "shader.h"
-#include "shader_program.h"
-#include "shader_source.h"
 #include "opengl/opengl.h"
 #include "opengl/texture.h"
 #include "opengl/frame_buffer.h"
+#include "opengl/shader_program.h"
+#include "opengl/shader_source.h"
 
 
 namespace Rays
